+++ material/src/modules/debugDataCopyClick.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from '@angular/core';
-import {DebugDataModule} from '@anglr/common';
+import {DebugDataComponent} from '@anglr/common';
 
 import {DebugDataCopyClickDirective} from '../directives/debugDataCopyClick/debugDataCopyClick.directive';
 
@@ -10,15 +10,13 @@ import {DebugDataCopyClickDirective} from '../directives/debugDataCopyClick/debu
 {
     imports:
     [
-    ],
-    declarations:
-    [
-        DebugDataCopyClickDirective
+        DebugDataComponent,
+        DebugDataCopyClickDirective,
     ],
     exports:
     [
         DebugDataCopyClickDirective,
-        DebugDataModule,
+        DebugDataComponent,
     ]
 })
 export class DebugDataCopyClickModule