+++ src/pipes/index.ts
@@ -6,4 +6,5 @@ export * from './isPresent/isPresent.pipe';
 export * from './keys/keys.pipe';
 export * from './localize/localize.pipe';
 export * from './mergeCssClasses/mergeCssClasses.pipe';
+export * from './trimText/trimText.pipe';
 export * from './urlEncode/urlEncode.pipe';